backend/requirements.txt
bcrypt==3.1.7
Django==4.0.6
django-cors-headers==3.13.0
django-pdb==0.6.2
djangorestframework==3.14.0
drf-yasg==1.21.7
Pillow==7.0.0
psycopg2==2.9.3
python-dotenv==0.21.0
requests==2.22.0
